refactor: domain 폴더 속 slack 폴더 제거
