+++ slack-service/src/main/java/com/monkey/slackservice/domain/repository/SlackRepository.java
@@ -1,0 +1,8 @@
+package com.monkey.slackservice.domain.repository;
+
+import com.monkey.slackservice.domain.entity.SlackEntity;
+
+public interface SlackRepository {
+    SlackEntity save(SlackEntity slackEntity);
+    long count();
+}
