--- slack-service/src/main/java/com/monkey/slackservice/domain/slack/repository/SlackRepository.java
@@ -1,8 +1,0 @@
-package com.monkey.slackservice.domain.slack.repository;
-
-import com.monkey.slackservice.domain.slack.entity.SlackEntity;
-
-public interface SlackRepository {
-    SlackEntity save(SlackEntity slackEntity);
-    long count();
-}
