+++ slack-service/src/main/java/com/monkey/slackservice/infrastructure/persistence/slack/SlackJpaRepository.java
@@ -1,6 +1,6 @@
 package com.monkey.slackservice.infrastructure.persistence.slack;
 
-import com.monkey.slackservice.domain.slack.entity.SlackEntity;
+import com.monkey.slackservice.domain.entity.SlackEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
